Fix pyright issues
